save commit

+++ src/main/java/com/cml/filestorage/service/impl/FileServiceImpl.java
@@ -1,10 +1,14 @@
 package com.cml.filestorage.service.impl;
 
+import com.cml.filestorage.exception.FileDoesNotExistsException;
+import com.cml.filestorage.exception.InvalidInputException;
+import com.cml.filestorage.exception.TagDoesNotExistsException;
 import com.cml.filestorage.model.File;
 import com.cml.filestorage.repository.ElasticFileRepository;
 import com.cml.filestorage.service.FileService;
 import org.springframework.stereotype.Service;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FileServiceImpl implements FileService {
@@ -17,29 +21,53 @@ public class FileServiceImpl implements FileService {
 
     @Override
     public File save(File file) {
+        if (file.getSize() <= 0 || file.getName().length() == 0) {
+            throw new InvalidInputException("File has incorrect name or size");
+        }
         return fileRepository.save(file);
     }
 
     @Override
-    public File getFileById(Long id) {
-        return fileRepository.getFileById(id);
+    public void deleteById(String id) {
+        Optional<File> fileOptional = fileRepository.findById(id);
+        if (fileOptional.isPresent()) {
+            fileRepository.deleteById(id);
+        } else {
+            throw new FileDoesNotExistsException("file not found");
+        }
     }
 
     @Override
-    public void deleteById(Long id) {
-        fileRepository.deleteById(id);
+    public File assignTags(String id, List<String> tagList) {
+        Optional<File> fileOptional = fileRepository.findById(id);
+        File file;
+        if (fileOptional.isPresent()) {
+            file = fileOptional.get();
+            file.getTagList().addAll(tagList);
+            fileRepository.deleteById(id);
+            fileRepository.save(file);
+            return file;
+        }
+        throw new FileDoesNotExistsException("file not found");
     }
 
     @Override
-    public File assignTags(File file, List<String> tagList) {
-        file.getTagList().addAll(tagList);
-        return fileRepository.update(file);
-    }
+    public File removeTags(String id, List<String> tagList) {
+        Optional<File> fileOptional = fileRepository.findById(id);
+        File file;
+        if (fileOptional.isPresent()) {
+            file = fileOptional.get();
+            if (file.getTagList().containsAll(tagList)) {
+                file.getTagList().removeAll(tagList);
+                fileRepository.deleteById(id);
+                fileRepository.save(file);
+                return file;
+            } else {
+                throw new TagDoesNotExistsException("tag not found on file");
+            }
 
-    @Override
-    public File removeTags(File file, List<String> tagList) {
-        file.getTagList().removeAll(tagList);
-        return fileRepository.update(file);
+        }
+        throw new FileDoesNotExistsException("file not found");
     }
 
     @Override
